Removed actions from index files
